requirements.txt
python-dotenv
numpy
pandas
scikit-learn
pymongo
certifi
pymongo[srv]==3.11
dill
pyaml
mlflow
# -e .